requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
email-validator
itsdangerous
psycopg2-binary
